fix: fix building static bundle
