+++ src/components/pages/LandingPage/sections/installation-prompt.tsx
